1.9.0

+++ Seewo/main.pyw
@@ -1,22 +1,15 @@
-v = "2.0.0"
-import filecmp
-import glob
+v = "2.1.0"
+
+import bs4
 import os
-import re
+import requests
 import shutil
-import stat
 import sys
 import time
 import webbrowser
-import winreg
 from tkinter import *
 from tkinter import ttk
 from tkinter.ttk import Separator
-
-import bs4
-import requests
-import send2trash
-import winshell
 
 date = time.strftime("%Y-%m-%d")
 # 初始化
@@ -41,25 +34,21 @@ tk.wm_iconbitmap("ico.ico")
 
 
 def pc_remove(d, name):
-    a = [k for (k, v) in d.items() if v == name]
-    for i in a:
-        del d[i]
+    for i in [k for (k, v) in d.items() if v == name]: del d[i]
 
 
 def remove_if_in(d, name):
     a = []
     for i in d.keys():
-        if name in i:
-            a.append(i)
-    for i in a:
-        del d[i]
+        if name in i: a.append(i)
+    for i in a: del d[i]
 
 
 def repeat_clear(path):
+    import filecmp, glob
     file_lst = []
     for i in glob.glob(path + '/**/*', recursive=True):
-        if os.path.isfile(i):
-            file_lst.append(i)
+        if os.path.isfile(i): file_lst.append(i)
     for x in file_lst:
         for y in file_lst:
             if x != y and os.path.exists(x) and os.path.exists(y):
@@ -71,10 +60,10 @@ def repeat_clear(path):
 
 
 def move_files(old, new):
+    import stat
     list2 = []
     list3 = os.walk(old)
-    for i in list3:
-        list2.append(i)
+    for i in list3: list2.append(i)
     try:
         list3 = list2[0][2]
     except:
@@ -86,134 +75,106 @@ def move_files(old, new):
     mp3 = []
     zip = []
     for i in list3:
-        if ".ppt" in i and os.path.exists(os.path.join(old + i)):
-            ppt.append(i)
-        if (".doc" in i or ".txt" in i or ".pdf" in i) and os.path.exists(os.path.join(old + i)):
-            doc.append(i)
-        if ".xls" in i and os.path.exists(os.path.join(old + i)):
-            xls.append(i)
-        if (
-                ".png" in i or ".jpg" in i or ".jpeg" in i or ".webp" in i or ".JPG" in i or ".PNG" in i or ".gif" in i) and os.path.exists(
-            os.path.join(old + i)):
-            img.append(i)
-        if ".mp" in i and os.path.exists(os.path.join(old + i)):
-            mp3.append(i)
-        if (".zip" in i or ".rar" in i or ".7z" in i) and os.path.exists(os.path.join(old + i)):
-            zip.append(i)
+        for j in [".ppt", ".PPT"]:
+            if j in i and os.path.exists(os.path.join(old + i)): ppt.append(i)
+        for j in [".doc", ".DOC", ".txt", ".TXT", ".pdf", ".PDF"]:
+            if j in i and os.path.exists(os.path.join(old + i)): doc.append(i)
+        for j in [".xls", ".XLS"]:
+            if j in i and os.path.exists(os.path.join(old + i)): xls.append(i)
+        for j in [".png", ".PNG", ".jpg", ".JPG", ".jpeg", ".JPEG", ".webp", ".WEBP", ".gif", ".GIF"]:
+            if j in i and os.path.exists(os.path.join(old + i)): img.append(i)
+        for j in [".mp", ".MP"]:
+            if j in i and os.path.exists(os.path.join(old + i)): mp3.append(i)
+        for j in [".zip", ".ZIP", ".rar", ".RAR", ".7z", ".7Z"]:
+            if j in i and os.path.exists(os.path.join(old + i)): zip.append(i)
     for i in range(len(ppt)):
         if os.path.exists(new + "PPT/" + ppt[i]):
             j = 1
-            while os.path.exists(
-                    new + "PPT/" + ppt[i][:ppt[i].rfind(".")] + "(" + str(j) + ")" + ppt[i][ppt[i].rfind("."):]):
-                j = j + 1
+            while os.path.exists(new + "PPT/" + ppt[i][:ppt[i].rfind(".")] + "(" + str(j) + ")" + ppt[i][ppt[i].rfind("."):]): j = j + 1
             ppt[i] = ppt[i] + "(" + str(j) + ")"
     for i in range(len(doc)):
         if os.path.exists(new + "文档/" + doc[i]):
             j = 1
-            while os.path.exists(
-                    new + "文档/" + doc[i][:doc[i].rfind(".")] + "(" + str(j) + ")" + doc[i][doc[i].rfind("."):]):
-                j = j + 1
+            while os.path.exists(new + "文档/" + doc[i][:doc[i].rfind(".")] + "(" + str(j) + ")" + doc[i][doc[i].rfind("."):]): j = j + 1
             doc[i] = doc[i] + "(" + str(j) + ")"
     for i in range(len(xls)):
         if os.path.exists(new + "表格/" + xls[i]):
             j = 1
-            while os.path.exists(
-                    new + "表格/" + xls[i][:xls[i].rfind(".")] + "(" + str(j) + ")" + xls[i][xls[i].rfind("."):]):
-                j = j + 1
+            while os.path.exists(new + "表格/" + xls[i][:xls[i].rfind(".")] + "(" + str(j) + ")" + xls[i][xls[i].rfind("."):]): j = j + 1
             xls[i] = xls[i] + "(" + str(j) + ")"
     for i in range(len(img)):
         if os.path.exists(new + "图片/" + img[i]):
             j = 1
-            while os.path.exists(
-                    new + "图片/" + img[i][:img[i].rfind(".")] + "(" + str(j) + ")" + img[i][img[i].rfind("."):]):
-                j = j + 1
+            while os.path.exists(new + "图片/" + img[i][:img[i].rfind(".")] + "(" + str(j) + ")" + img[i][img[i].rfind("."):]): j = j + 1
             img[i] = img[i] + "(" + str(j) + ")"
     for i in range(len(mp3)):
         if os.path.exists(new + "音视频/" + mp3[i]):
             j = 1
-            while os.path.exists(
-                    new + "音视频/" + mp3[i][:mp3[i].rfind(".")] + "(" + str(j) + ")" + mp3[i][mp3[i].rfind("."):]):
-                j = j + 1
+            while os.path.exists(new + "音视频/" + mp3[i][:mp3[i].rfind(".")] + "(" + str(j) + ")" + mp3[i][mp3[i].rfind("."):]): j = j + 1
             mp3[i] = mp3[i] + "(" + str(j) + ")"
     for i in range(len(zip)):
         if os.path.exists(new + "压缩包/" + zip[i]):
             j = 1
-            while os.path.exists(
-                    new + "压缩包/" + zip[i][:zip[i].rfind(".")] + "(" + str(j) + ")" + zip[i][zip[i].rfind("."):]):
-                j = j + 1
+            while os.path.exists(new + "压缩包/" + zip[i][:zip[i].rfind(".")] + "(" + str(j) + ")" + zip[i][zip[i].rfind("."):]): j = j + 1
             zip[i] = zip[i] + "(" + str(j) + ")"
-    if not os.path.exists(new + "PPT/"):
-        os.makedirs(new + "PPT/")
-    if not os.path.exists(new + "表格/"):
-        os.makedirs(new + "表格/")
-    if not os.path.exists(new + "文档/"):
-        os.makedirs(new + "文档/")
-    if not os.path.exists(new + "图片/"):
-        os.makedirs(new + "图片/")
-    if not os.path.exists(new + "音视频/"):
-        os.makedirs(new + "音视频/")
-    if not os.path.exists(new + "压缩包/"):
-        os.makedirs(new + "压缩包/")
-    if not os.path.exists(new + "文件夹/"):
-        os.makedirs(new + "文件夹/")
+    if not os.path.exists(new + "PPT/"): os.makedirs(new + "PPT/")
+    if not os.path.exists(new + "表格/"): os.makedirs(new + "表格/")
+    if not os.path.exists(new + "文档/"): os.makedirs(new + "文档/")
+    if not os.path.exists(new + "图片/"): os.makedirs(new + "图片/")
+    if not os.path.exists(new + "音视频/"): os.makedirs(new + "音视频/")
+    if not os.path.exists(new + "压缩包/"): os.makedirs(new + "压缩包/")
+    if not os.path.exists(new + "文件夹/"): os.makedirs(new + "文件夹/")
+
     for i in ppt:
         try:
             os.chmod(old + i, stat.S_IWRITE)
             shutil.move(old + i, new + "PPT/" + i)
         except:
             os.chmod(old + i[:i.rfind("(")], stat.S_IWRITE)
-            shutil.move(old + i[:i.rfind("(")],
-                        new + "PPT/" + i[:i.rfind(".")] + i[i.rfind("("):] + i[i.rfind("."):i.rfind("(")])
+            shutil.move(old + i[:i.rfind("(")], new + "PPT/" + i[:i.rfind(".")] + i[i.rfind("("):] + i[i.rfind("."):i.rfind("(")])
     for i in doc:
         try:
             os.chmod(old + i, stat.S_IWRITE)
             shutil.move(old + i, new + "文档/" + i)
         except:
             os.chmod(old + i[:i.rfind("(")], stat.S_IWRITE)
-            shutil.move(old + i[:i.rfind("(")],
-                        new + "文档/" + i[:i.rfind(".")] + i[i.rfind("("):] + i[i.rfind("."):i.rfind("(")])
+            shutil.move(old + i[:i.rfind("(")], new + "文档/" + i[:i.rfind(".")] + i[i.rfind("("):] + i[i.rfind("."):i.rfind("(")])
     for i in xls:
         try:
             os.chmod(old + i, stat.S_IWRITE)
             shutil.move(old + i, new + "表格/" + i)
         except:
             os.chmod(old + i[:i.rfind("(")], stat.S_IWRITE)
-            shutil.move(old + i[:i.rfind("(")],
-                        new + "表格/" + i[:i.rfind(".")] + i[i.rfind("("):] + i[i.rfind("."):i.rfind("(")])
+            shutil.move(old + i[:i.rfind("(")], new + "表格/" + i[:i.rfind(".")] + i[i.rfind("("):] + i[i.rfind("."):i.rfind("(")])
     for i in img:
         try:
             os.chmod(old + i, stat.S_IWRITE)
             shutil.move(old + i, new + "图片/" + i)
         except:
             os.chmod(old + i[:i.rfind("(")], stat.S_IWRITE)
-            shutil.move(old + i[:i.rfind("(")],
-                        new + "图片/" + i[:i.rfind(".")] + i[i.rfind("("):] + i[i.rfind("."):i.rfind("(")])
+            shutil.move(old + i[:i.rfind("(")], new + "图片/" + i[:i.rfind(".")] + i[i.rfind("("):] + i[i.rfind("."):i.rfind("(")])
     for i in mp3:
         try:
             os.chmod(old + i, stat.S_IWRITE)
             shutil.move(old + i, new + "音视频/" + i)
         except:
             os.chmod(old + i[:i.rfind("(")], stat.S_IWRITE)
-            shutil.move(old + i[:i.rfind("(")],
-                        new + "音视频/" + i[:i.rfind(".")] + i[i.rfind("("):] + i[i.rfind("."):i.rfind("(")])
+            shutil.move(old + i[:i.rfind("(")], new + "音视频/" + i[:i.rfind(".")] + i[i.rfind("("):] + i[i.rfind("."):i.rfind("(")])
     for i in zip:
         try:
             os.chmod(old + i, stat.S_IWRITE)
             shutil.move(old + i, new + "压缩包/" + i)
         except:
             os.chmod(old + i[:i.rfind("(")], stat.S_IWRITE)
-            shutil.move(old + i[:i.rfind("(")],
-                        new + "压缩包/" + i[:i.rfind(".")] + i[i.rfind("("):] + i[i.rfind("."):i.rfind("(")])
+            shutil.move(old + i[:i.rfind("(")], new + "压缩包/" + i[:i.rfind(".")] + i[i.rfind("("):] + i[i.rfind("."):i.rfind("(")])
     list3 = list2[0][1]
     fold = []
     for i in list3:
-        if i != "软件":
-            fold.append(i)
+        if i != "软件": fold.append(i)
     for i in range(len(fold)):
         if os.path.exists(new + "文件夹/" + fold[i]):
             j = 1
-            while os.path.exists(new + "文件夹/" + fold[i] + "(" + str(j) + ")"):
-                j = j + 1
+            while os.path.exists(new + "文件夹/" + fold[i] + "(" + str(j) + ")"): j = j + 1
             fold[i] = fold[i] + "(" + str(j) + ")"
     for i in fold:
         try:
@@ -227,20 +188,18 @@ def move_files(old, new):
 
 def b0():
     print("检查更新")
+    import re
     link = "https://ianzb.github.io/server.github.io/Seewo/"
-    res = requests.get(
-        link + "seewo.html")
+    res = requests.get(link + "seewo.html")
     res.encoding = "UTF-8"
     soup = bs4.BeautifulSoup(res.text, "lxml")
     data = soup.find_all(name="div", text=re.compile("."))
-    for i in range(len(data)):
-        data[i] = str(data[i]).replace("<div>", "").replace("</div>", "").replace(r"\r", "").replace(r"\n", "").strip()
+    for i in range(len(data)): data[i] = str(data[i]).replace("<div>", "").replace("</div>", "").replace(r"\r", "").replace(r"\n", "").strip()
     for i in range(len(data)):
         response1 = requests.get(link + data[i])
         response1.encoding = "UTF-8"
         main = response1.content
-        with open(data[i], "wb") as file:
-            file.write(main)
+        with open(data[i], "wb") as file: file.write(main)
         print(data[i])
     os.popen("main.pyw")
     sys.exit()
@@ -260,20 +219,16 @@ def b101():
     response.encoding = "UTF-8"
     soup = bs4.BeautifulSoup(response.text, "lxml")
     data1 = soup.find_all(name="td")
-    for n in data1:
-        a.append(n.text)
+    for n in data1: a.append(n.text)
     for i in range(len(a)):
-        if i % 3 != 2:
-            b.append(a[i])
+        if i % 3 != 2: b.append(a[i])
     for i in range(len(b)):
-        if i % 2 == 0:
-            v[b[i]] = b[i + 1]
+        if i % 2 == 0: v[b[i]] = b[i + 1]
     pc_remove(v, "")
-    for c in ["内部", "Windows", "macOS", "Linux", "即将到来", "ChromeOS", "PlayStation", "Nintendo", "Xbox", "Steam", "demo", "教育版（iOS）", "Minecraft Dungeons（启动器版）", "战斗测试"]:
-        remove_if_in(v, c)
+    list = ["内部", "Windows", "macOS", "Linux", "即将到来", "ChromeOS", "PlayStation", "Nintendo", "Xbox", "Steam", "demo", "教育版（iOS）", "Minecraft Dungeons（启动器版）", "战斗测试"]
+    for c in list: remove_if_in(v, c)
     with open("mc.txt", "w", encoding="utf-8") as file:
-        for (k, v) in v.items():
-            file.write(k + "版本：" + v + "\n")
+        for (k, v) in v.items(): file.write(k + "版本：" + v + "\n")
     os.popen("mc.txt")
     time.sleep(1)
     os.remove("mc.txt")
@@ -305,21 +260,21 @@ def b4():
 
 def b5():
     print("清理希沃视频展台文件")
+    import send2trash
     try:
         list = os.walk(r"D:/EasiCameraPhoto")
         list2 = []
-        for i in list:
-            list2.append(i)
+        for i in list: list2.append(i)
         list = list2[0][1]
         for i in list:
-            if i != date and os.path.exists(r"D:/EasiCameraPhoto/" + i):
-                send2trash.send2trash(os.path.join(r"D:\EasiCameraPhoto\ "[:-1] + i))
+            if i != date and os.path.exists(r"D:/EasiCameraPhoto/" + i): send2trash.send2trash(os.path.join(r"D:\EasiCameraPhoto\ "[:-1] + i))
     except:
         pass
 
 
 def b6():
     print("清理回收站")
+    import winshell
     try:
         winshell.recycle_bin().empty(confirm=False, show_progress=False, sound=False)
     except:
@@ -328,6 +283,7 @@ def b6():
 
 def b7():
     print("整理桌面文件")
+    import winreg
     key = winreg.OpenKey(winreg.HKEY_CURRENT_USER, r'Software\Microsoft\Windows\CurrentVersion\Explorer\Shell Folders')
     path = winreg.QueryValueEx(key, "Desktop")[0] + r"\ "[0:-1]
     move_files(path, "D:/文件/")
@@ -345,21 +301,16 @@ def b8():
     try:
         list = []
         list2 = []
-        for i in os.walk("D:/WeChat Files/WeChat Files/"):
-            list.append(i)
+        for i in os.walk("D:/WeChat Files/WeChat Files/"): list.append(i)
         for i in list[0][1]:
-            if os.path.exists(os.path.join("D:/WeChat Files/WeChat Files/", i, "FileStorage\File")):
-                list2.append(os.path.join("D:/WeChat Files/WeChat Files/", i, "FileStorage\File"))
+            if os.path.exists(os.path.join("D:/WeChat Files/WeChat Files/", i, "FileStorage\File")): list2.append(os.path.join("D:/WeChat Files/WeChat Files/", i, "FileStorage\File"))
         list = []
         list3 = []
         for i in range(len(list2)):
-            for j in os.walk(list2[i]):
-                list.append(j)
-            for k in list[0][1]:
-                list3.append(os.path.join(list2[i], k))
+            for j in os.walk(list2[i]): list.append(j)
+            for k in list[0][1]: list3.append(os.path.join(list2[i], k))
         list = list3
-        for i in list:
-            move_files(i + "/", "D:/文件/")
+        for i in list: move_files(i + "/", "D:/文件/")
         repeat_clear("D:/文件/PPT/")
         repeat_clear("D:/文件/表格/")
         repeat_clear("D:/文件/图片/")
@@ -373,6 +324,7 @@ def b8():
 
 def b9():
     print("清理整理文件")
+    import send2trash
     try:
         send2trash.send2trash(r"D:\文件")
     except:
@@ -383,10 +335,8 @@ def b10():
     print("清理系统缓存")
     list = []
     list1 = os.walk(os.getenv("TEMP"))
-    for i in list1:
-        list.append(i)
-    if list != []:
-        list1 = list[0][1]
+    for i in list1: list.append(i)
+    if list != []: list1 = list[0][1]
     list2 = list[0][2]
     for i in list1:
         try:
@@ -414,57 +364,48 @@ def b12():
     time.sleep(0.1)
     os.popen("C:\Program Files (x86)\Seewo\PPTService\Main\PPTService.exe")
     print("清理希沃视频展台文件")
+    import send2trash
     try:
         list = os.walk(r"D:/EasiCameraPhoto")
         list2 = []
-        for i in list:
-            list2.append(i)
+        for i in list: list2.append(i)
         list = list2[0][1]
         for i in list:
-            if i != date and os.path.exists(r"D:/EasiCameraPhoto/" + i):
-                send2trash.send2trash(
-                    os.path.join(r"D:\EasiCameraPhoto\ "[:-1] + i))
+            if i != date and os.path.exists(r"D:/EasiCameraPhoto/" + i): send2trash.send2trash(os.path.join(r"D:\EasiCameraPhoto\ "[:-1] + i))
     except:
         pass
     print("清理回收站")
+    import winshell
     try:
         winshell.recycle_bin().empty(confirm=False, show_progress=False, sound=False)
     except:
         pass
     print("整理桌面文件")
-    key = winreg.OpenKey(winreg.HKEY_CURRENT_USER,
-                         r'Software\Microsoft\Windows\CurrentVersion\Explorer\Shell Folders')
+    import winreg
+    key = winreg.OpenKey(winreg.HKEY_CURRENT_USER, r'Software\Microsoft\Windows\CurrentVersion\Explorer\Shell Folders')
     path = winreg.QueryValueEx(key, "Desktop")[0] + r"\ "[0:-1]
     move_files(path, "D:/文件/")
     print("整理微信文件")
     try:
         list = []
         list2 = []
-        for i in os.walk("D:/WeChat Files/WeChat Files/"):
-            list.append(i)
+        for i in os.walk("D:/WeChat Files/WeChat Files/"): list.append(i)
         for i in list[0][1]:
-            if os.path.exists(os.path.join("D:/WeChat Files/WeChat Files/", i, "FileStorage\File")):
-                list2.append(
-                    os.path.join("D:/WeChat Files/WeChat Files/", i, "FileStorage\File"))
+            if os.path.exists(os.path.join("D:/WeChat Files/WeChat Files/", i, "FileStorage\File")): list2.append(os.path.join("D:/WeChat Files/WeChat Files/", i, "FileStorage\File"))
         list = []
         list3 = []
         for i in range(len(list2)):
-            for j in os.walk(list2[i]):
-                list.append(j)
-            for k in list[0][1]:
-                list3.append(os.path.join(list2[i], k))
+            for j in os.walk(list2[i]): list.append(j)
+            for k in list[0][1]: list3.append(os.path.join(list2[i], k))
         list = list3
-        for i in list:
-            move_files(i + "/", "D:/文件/")
+        for i in list: move_files(i + "/", "D:/文件/")
     except:
         pass
     print("清理系统缓存")
     list = []
     list1 = os.walk(os.getenv("TEMP"))
-    for i in list1:
-        list.append(i)
-    if list != []:
-        list1 = list[0][1]
+    for i in list1: list.append(i)
+    if list != []: list1 = list[0][1]
     list2 = list[0][2]
     for i in list1:
         try:
